+++ src/main/java/com/example/shimpyo/domain/auth/service/MailService.java
@@ -2,6 +2,7 @@ package com.example.shimpyo.domain.auth.service;
 
 import com.example.shimpyo.domain.auth.dto.MailVerifyDto;
 import com.example.shimpyo.domain.auth.dto.MailCodeSendDto;
+import com.example.shimpyo.domain.auth.entity.UserAuth;
 import com.example.shimpyo.domain.user.repository.UserRepository;
 import com.example.shimpyo.global.BaseException;
 import jakarta.mail.MessagingException;
@@ -34,12 +35,6 @@ public class MailService {
     @Qualifier("1")
     private final RedisTemplate<String, Object> redisTemplate;
     private final UserRepository userRepository;
-
-    private static final String LETTERS = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ";
-    private static final String NUMBERS = "0123456789";
-    private static final String SPECIALS = "~!@#$%^&*";
-    private static final String ALL = LETTERS + NUMBERS + SPECIALS;
-    private static final SecureRandom random = new SecureRandom();
     
     // [#MOO4] 메일 전송 시작 
     public void authEmail(MailCodeSendDto dto) {
@@ -103,9 +98,10 @@ public class MailService {
     // [#MOO5] 메일 정보와 인증 코드 일치 여부 판단 로직 끝
 
     // 회원 이메일로 임시 비밀번호 전송
-    public void sendResetPasswordMail(String email) throws MessagingException {
+    public void sendResetPasswordMail(String email, String tempPW) throws MessagingException {
+
         String subject = "ShimPyoSo Authorization";
-        String text = "임시 비밀번호는 " + generatePassword() + "입니다.";
+        String text = "임시 비밀번호는 " + tempPW + "입니다.";
 
         MimeMessage mimeMessage = mailSender.createMimeMessage();
         MimeMessageHelper helper = new MimeMessageHelper(mimeMessage, true, "utf-8");
@@ -116,21 +112,5 @@ public class MailService {
         mailSender.send(mimeMessage);
     }
 
-    private static String generatePassword() {
-        List<Character> passwordChars = new ArrayList<>();
 
-        passwordChars.add(LETTERS.charAt(random.nextInt(LETTERS.length())));
-        passwordChars.add(NUMBERS.charAt(random.nextInt(NUMBERS.length())));
-        passwordChars.add(SPECIALS.charAt(random.nextInt(SPECIALS.length())));
-
-        for (int i = 3; i < 8; i++) {
-            passwordChars.add(ALL.charAt(random.nextInt(ALL.length())));
-        }
-        Collections.shuffle(passwordChars);
-        StringBuilder password = new StringBuilder();
-        for (char ch : passwordChars) {
-            password.append(ch);
-        }
-        return password.toString();
-    }
 }
